Implementado a classe DatabaseIntegrityViolationException como exception personalizada

+++ src/main/java/com/walker/wncommercejavaspring/service/ProductService.java
@@ -3,14 +3,16 @@ package com.walker.wncommercejavaspring.service;
 import com.walker.wncommercejavaspring.dto.ProductDto;
 import com.walker.wncommercejavaspring.model.Product;
 import com.walker.wncommercejavaspring.repository.ProductRepository;
+import com.walker.wncommercejavaspring.service.exception.DatabaseIntegrityViolationException;
 import com.walker.wncommercejavaspring.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -40,15 +42,27 @@ public class ProductService {
 
     @Transactional
     public ProductDto update(Long id, ProductDto productDto) {
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDto, product);
-        productRepository.save(product);
-        return new ProductDto(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDto, product);
+            productRepository.save(product);
+            return new ProductDto(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseIntegrityViolationException("Violação da integridade referencial!");
+        }
+
     }
